+++ History.md
@@ -2,6 +2,7 @@
 
 * Features
   * Add pumactl `thread-backtraces` command to print thread backtraces (#2053)
+  * Configuration: `environment` is read from `RAILS_ENV`, if `RACK_ENV` can't be found (#2022)
 
 * Bugfixes
   * Your bugfix goes here (#Github Number)
@@ -12,7 +13,9 @@
 * Features
   * Strip whitespace at end of HTTP headers (#2010)
   * Optimize HTTP parser for JRuby (#2012)
+  * Add SSL support for the control app (#2046)
   * Add SSL support for the control app and cli (#2046, #2052)
+
 
 * Bugfixes
   * Fix Errno::EINVAL when SSL is enabled and browser rejects cert (#1564)